CarouselPagination is shared
