+++ src/ui/components/CarouselPagination/CarouselPagination.tsx
@@ -1,0 +1,3 @@
+import { CarouselPaginationBase } from 'ui/components/CarouselPagination/CarouselPaginationBase'
+
+export const CarouselPagination = CarouselPaginationBase
