+++ src/ui/components/CarouselPagination/CarouselPaginationBase.tsx
@@ -1,20 +1,14 @@
 import React, { FunctionComponent } from 'react'
-import { SharedValue } from 'react-native-reanimated'
 import styled from 'styled-components/native'
 import { v4 as uuidv4 } from 'uuid'
 
 import { CarouselDot } from 'ui/CarouselDot/CarouselDot'
+import { CarouselPaginationProps } from 'ui/components/CarouselPagination/types'
 import { ViewGap } from 'ui/components/ViewGap/ViewGap'
 
-type Props = {
-  progressValue: SharedValue<number>
-  offerImages: string[]
-  gap: number
-}
-
-export const OfferImageCarouselDots: FunctionComponent<Props> = ({
+export const CarouselPaginationBase: FunctionComponent<CarouselPaginationProps> = ({
   progressValue,
-  offerImages,
+  elementsCount,
   gap,
 }) => {
   const carouselDotId = uuidv4()
@@ -23,7 +17,7 @@ export const OfferImageCarouselDots: FunctionComponent<Props> = ({
 
   return (
     <Container gap={gap}>
-      {offerImages.map((_, index) => (
+      {Array.from({ length: elementsCount }).map((_, index) => (
         <CarouselDot animValue={progressValue} index={index} key={index + carouselDotId} />
       ))}
     </Container>
